server/requirements.txt
fastapi
uvicorn[standard]
aiortc
numpy
opencv-python-headless
requests
python-multipart
torch
torchvision
pillow
qrcode[pil]
ultralytics
scipy
